+++ volmdlr/shapes.py
@@ -7,7 +7,7 @@ import warnings
 import zlib
 from io import BytesIO
 from itertools import chain, product
-from typing import Iterable, List, Tuple, Union, Optional, Any, Dict, overload
+from typing import Iterable, List, Tuple, Union, Optional, Any, Dict, overload, Literal, cast as tcast
 
 import matplotlib.pyplot as plt
 import networkx as nx
@@ -32,6 +32,8 @@ from OCP.BOPAlgo import BOPAlgo_GlueEnum, BOPAlgo_PaveFiller
 from OCP.TopTools import TopTools_ListOfShape
 from OCP.ShapeFix import ShapeFix_Solid
 from OCP.BRepTools import BRepTools
+from OCP.TopTools import TopTools_IndexedMapOfShape
+from OCP.TopExp import TopExp
 
 import volmdlr.core_compiled
 from volmdlr import curves, display, edges, surfaces, wires, geometry, faces as vm_faces
@@ -42,6 +44,21 @@ from volmdlr.utils.mesh_helpers import perform_decimation
 
 import OCP.TopAbs as topabs  # Topology type enum
 
+shape_LUT = {
+    topabs.TopAbs_VERTEX: "Vertex",
+    topabs.TopAbs_EDGE: "Edge",
+    topabs.TopAbs_WIRE: "Wire",
+    topabs.TopAbs_FACE: "Face",
+    topabs.TopAbs_SHELL: "Shell",
+    topabs.TopAbs_SOLID: "Solid",
+    topabs.TopAbs_COMPSOLID: "CompSolid",
+    topabs.TopAbs_COMPOUND: "Compound",
+}
+
+inverse_shape_LUT = {v: k for k, v in shape_LUT.items()}
+
+Shapes = Literal[
+    "Vertex", "Edge", "Wire", "Face", "Shell", "Solid", "CompSolid", "Compound"]
 
 # pylint: disable=no-name-in-module,invalid-name,unused-import,wrong-import-order
 
@@ -88,6 +105,7 @@ class Shape(PhysicalObject):
     def __init__(self, obj: TopoDS_Shape, name: str = ""):
         self.wrapped = downcast(obj)
         self.label = name
+        self._bbox = None
         PhysicalObject.__init__(self, name=name)
 
     @classmethod
@@ -112,10 +130,54 @@ class Shape(PhysicalObject):
 
         return tr
 
+    @staticmethod
+    def _entities(obj, topo_type: Shapes) -> Iterable[TopoDS_Shape]:
+        shape_set = TopTools_IndexedMapOfShape()
+        TopExp.MapShapes_s(obj, inverse_shape_LUT[topo_type], shape_set)
+
+        return tcast(Iterable[TopoDS_Shape], shape_set)
+
+    def _get_vertices(self):
+        return [downcast(i) for i in self._entities(obj=self.wrapped, topo_type="Vertex")]
+
+    def _get_edges(self):
+        return [downcast(i) for i in self._entities(obj=self.wrapped, topo_type="Edge") if
+                not BRep_Tool.Degenerated_s(TopoDS.Edge_s(i))]
+
+    def _get_faces(self):
+        return [downcast(i) for i in self._entities(obj=self.wrapped, topo_type="Face")]
+
+    def bounding_box(self):
+        if not self._bbox:
+            tol = 1e-2
+            bbox = Bnd_Box()
+
+            mesh = BRepMesh_IncrementalMesh(self.wrapped, tol, True)
+            mesh.Perform()
+
+            BRepBndLib.Add_s(self.wrapped, bbox, True)
+
+            xmin, ymin, zmin, xmax, ymax, zmax = bbox.Get()
+
+            self._bbox = volmdlr.core.BoundingBox(xmin=xmin, xmax=xmax, ymin=ymin, ymax=ymax, zmin=zmin, zmax=zmax)
+        return self._bbox
+
+    def volume(self):
+        """
+        Gets the Volume of a shape.
+
+        :return:
+        """
+        tol = 1e-6
+        prop = GProp_PGProps()
+        BRepGProp.VolumeProperties_s(self.wrapped, prop, tol)
+        return abs(prop.Mass())
+
 
 class Shell(Shape):
     """
     OCP shell wrapped.
+
     """
 
     wrapped: TopoDS_Shell
@@ -128,20 +190,47 @@ class Shell(Shape):
     def __init__(self, faces: List[TopoDS_Face], name: str = '') -> None:
         ...
 
-    def __init__(self, faces: List[vm_faces], name: str = ''):
-        self._faces = None
-        if isinstance(faces[0], vm_faces.Face3D):
-            obj = None
-            self._faces = faces
-        elif isinstance(faces[0], TopoDS_Face):
-            obj = None
-        else:
-            raise ValueError(
-                f"Provided faces value: {faces} is not a valid list of faces."
-            )
+    @overload
+    def __init__(self, faces: List[vm_faces.Face3D], name: str = '') -> None:
+        ...
 
-        self._bbox = None
+    def __init__(self, faces: List[vm_faces.Face3D] = None, name: str = '', obj=None):
+        self._faces = None
+        if faces:
+            obj = self._from_faces(faces)
+            if isinstance(faces[0], vm_faces.Face3D):
+                self._faces = faces
         Shape.__init__(self, obj, name=name)
+
+    @staticmethod
+    def _from_faces(faces):
+        """
+        Helper method to create a TopoDS_Shell from a list of faces.
+
+        :param faces: list of faces volmdlr or a list of faces TopoDS_Face.
+        :return: TopoDS_Shell object.
+        """
+        if isinstance(faces[0], vm_faces.Face3D):
+            faces = [face.to_ocp() for face in faces]
+
+        shell_builder = BRepBuilderAPI_Sewing()
+
+        for face in faces:
+            shell_builder.Add(face)
+
+        shell_builder.Perform()
+        return shell_builder.SewedShape()
+
+    @property
+    def faces(self):
+        if not self._faces:
+            pass
+            # self._faces = [from_ocp. for face in self._get_faces(self.wrapped)]
+        return self._faces
+
+    @faces.setter
+    def faces(self, faces):
+        self._faces = faces
 
 
 class Solid(Shape):